requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.1
aiolimiter==1.1.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Deprecated==1.2.14
deprecation==2.1.0
distro==1.9.0
eval_type_backport==0.2.0
exceptiongroup==1.2.2
fastapi==0.115.5
fonttools==4.55.0
frozenlist==1.5.0
google-auth==2.36.0
gotrue==2.10.0
griffe==1.5.1
groq==0.13.0
h11==0.14.0
h2==4.1.0
haversine==2.9.0
hpack==4.0.0
httpcore==1.0.6
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
importlib_resources==6.4.5
jiter==0.8.0
kiwisolver==1.4.7
limits==3.13.0
logfire-api==2.6.0
matplotlib==3.9.2
multidict==6.1.0
networkx==3.4.2
numpy==2.1.3
openai==1.56.1
packaging==24.2
pillow==11.0.0
postgrest==0.18.0
propcache==0.2.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.3
pydantic-ai==0.0.8
pydantic-ai-slim==0.0.8
pydantic_core==2.27.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.19
realtime==2.0.6
redis==5.2.0
requests==2.32.3
rsa==4.9
six==1.16.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.41.2
storage3==0.9.0
StrEnum==0.4.15
supabase==2.10.0
supafunc==0.7.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
websockets==13.1
wrapt==1.16.0
yarl==1.17.1
